requirements.txt
browserstack-local
jsonmerge
selenium==4.*
psutil
pytest
pytest-variables
pytest-selenium
browserstack-sdk
pytest-html==3.2.0
python-dotenv